部署测试版准备

+++ src/main/resources/db/migration/V12__Create_users_table_and_auth_table.sql
@@ -2,9 +2,9 @@
 DROP TABLE IF EXISTS `user_auth_rel`;
 
 CREATE TABLE `user_auth_rel` (
-                                 `id` int(11) NOT NULL AUTO_INCREMENT,
-                                 `user_id` int(11) DEFAULT NULL,
-                                 `auth_id` int(11) DEFAULT NULL,
+                                 `id` bigint(11) NOT NULL AUTO_INCREMENT,
+                                 `user_id` bigint(11) DEFAULT NULL,
+                                 `auth_id` bigint(11) DEFAULT NULL,
                                  `auth_type` varchar(20) DEFAULT NULL,
                                  PRIMARY KEY (`id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
@@ -12,7 +12,7 @@ CREATE TABLE `user_auth_rel` (
 DROP TABLE IF EXISTS `user_third_auth`;
 
 CREATE TABLE `user_third_auth` (
-                                   `auth_id` int(11) NOT NULL AUTO_INCREMENT COMMENT '自增id',
+                                   `auth_id` bigint(11) NOT NULL AUTO_INCREMENT COMMENT '自增id',
                                    `openid` int(11) DEFAULT NULL COMMENT '第三方用户唯一标识id',
                                    `login_type` varchar(40) DEFAULT NULL COMMENT '登陆方式如github,gitee,qq....',
                                    `access_token` varchar(1000) DEFAULT NULL COMMENT '访问openaoi的令牌',
@@ -25,7 +25,7 @@ CREATE TABLE `user_third_auth` (
 DROP TABLE IF EXISTS `users`;
 
 CREATE TABLE `users` (
-                         `user_id` int(11) NOT NULL AUTO_INCREMENT,
+                         `user_id` bigint(11) NOT NULL AUTO_INCREMENT,
                          `token` varchar(200) DEFAULT NULL,
                          `user_info` varchar(200) DEFAULT NULL COMMENT '。。。。',
                          `avatar_url` varchar(100) DEFAULT NULL COMMENT '头像',
@@ -36,7 +36,7 @@ CREATE TABLE `users` (
 DROP TABLE IF EXISTS `user_local_auth`;
 
 CREATE TABLE `user_local_auth` (
-                                   `auth_id` int(11) NOT NULL AUTO_INCREMENT,
+                                   `auth_id` bigint(11) NOT NULL AUTO_INCREMENT,
                                    `user_name` varchar(100) DEFAULT NULL,
                                    `user_password` varchar(100) DEFAULT NULL,
                                    `mobile` int(11) DEFAULT NULL,
